Factor out logging out all users.
