+++ realm/realm-library/src/syncIntegrationTest/java/io/realm/objectserver/utils/UserFactory.java
@@ -16,13 +16,25 @@
 
 package io.realm.objectserver.utils;
 
-import java.util.UUID;
+import android.os.Handler;
+import android.os.HandlerThread;
 
+import java.util.Map;
+import java.util.UUID;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.atomic.AtomicInteger;
+
+import io.realm.AuthenticationListener;
 import io.realm.Realm;
 import io.realm.RealmConfiguration;
 import io.realm.SyncCredentials;
+import io.realm.SyncManager;
 import io.realm.SyncUser;
+import io.realm.TestHelper;
 import io.realm.log.RealmLog;
+
+import static org.junit.Assert.fail;
+
 
 // Helper class to retrieve users with same IDs even in multi-processes.
 // Must be in `io.realm.objectserver` to work around package protected methods.
@@ -118,4 +130,54 @@ public class UserFactory {
         return instance;
     }
 
+    /**
+     * Blocking call that logs out all users
+     */
+    public static void logoutAllUsers() {
+        final CountDownLatch allUsersLoggedOut = new CountDownLatch(1);
+        final HandlerThread ht = new HandlerThread("LoggingOutUsersThread");
+        ht.start();
+        Handler handler = new Handler(ht.getLooper());
+        handler.post(new Runnable() {
+            @Override
+            public void run() {
+                final AtomicInteger usersLoggedOut = new AtomicInteger(0);
+                final int activeUsers = SyncUser.all().size();
+                final AuthenticationListener listener = new AuthenticationListener() {
+                    @Override
+                    public void loggedIn(SyncUser user) {
+                        SyncManager.removeAuthenticationListener(this);
+                        fail("User logged in while exiting test: " + user);
+                    }
+
+                    @Override
+                    public void loggedOut(SyncUser user) {
+                        if (usersLoggedOut.incrementAndGet() == activeUsers) {
+                            SyncManager.removeAuthenticationListener(this);
+                            allUsersLoggedOut.countDown();
+                        }
+                    }
+                };
+                SyncManager.addAuthenticationListener(listener);
+
+                Map<String, SyncUser> users = SyncUser.all();
+                if (users.isEmpty()) {
+                    SyncManager.removeAuthenticationListener(listener);
+                    allUsersLoggedOut.countDown();
+                } else {
+                    for (SyncUser user : users.values()) {
+                        user.logout();
+                        if (!user.getAuthenticationUrl().toString().contains("127.0.0.1")) {
+                            // For dummy users, calling `logout()` will never result in the
+                            // authentication listener to trigger since the URL doesn't exist.
+                            // For these cases, we manually trigger the listener.
+                            listener.loggedOut(user);
+                        }
+                    }
+                }
+            }
+        });
+        TestHelper.awaitOrFail(allUsersLoggedOut);
+        ht.quit();
+    }
 }
